* Added log_capture_text_to_file and place_capture_text_on_clipboard options in config_template.txt.  If enabled, they will append translations to translation_log.txt, or force it into your clipboard.  Disabled by default.  Can be set to "disabled", "full", "pre", & "post" * Oops, wasn't including the font for Punjabi before (lohit.punjabi.1.1.ttf, its license is also included) * Fonts are now user configurable per language, see fonts.txt file for directions * Added contributer credit to Meerkov in the readme * Fixed some compiler warnings

+++ source/HotKeyHandler.cpp
@@ -10,7 +10,7 @@ extern AudioHandle g_lastAudioHandle;
 
 void HandleWMHotkey(UINT message, WPARAM wParam, LPARAM lParam)
 {
-	int ID = wParam;
+	int ID = (int)wParam;
 	//LogMsg("Got hotkey %d, %d, ID: %d", message, lParam, ID);
 	
 	GetApp()->HandleHotKeyPushed(GetApp()->m_hotKeyHandler.GetHotKeyByID(ID));
